+++ src/main/java/com/elmakers/mine/bukkit/spell/BaseSpell.java
@@ -2253,7 +2253,7 @@ public abstract class BaseSpell implements MageSpell, Cloneable {
             // Reward SP
             Wand wand = context.getWand();
             WandUpgradePath path = wand == null ? null : wand.getPath();
-            if (earns > 0 && wand != null && path != null && path.earnsSP() && controller.isSPEnabled() && !mage.isAtMaxSkillPoints()) {
+            if (earns > 0 && wand != null && path != null && path.earnsSP() && controller.isSPEnabled() && controller.isSPEarnEnabled() && !mage.isAtMaxSkillPoints()) {
                 long now = System.currentTimeMillis();
                 int scaledEarn = earns;
                 if (lastEarn > 0 && earnCooldown > 0 && now < lastEarn + earnCooldown) {